Filesystem: partial x64 port

+++ src/xrCore/file_stream_reader.cpp
@@ -4,13 +4,13 @@
 #include <fcntl.h>
 #endif
 
-void CFileStreamReader::construct(pcstr file_name, const u32& window_size)
+void CFileStreamReader::construct(pcstr file_name, const size_t& window_size)
 {
 #if defined(WINDOWS)
     m_file_handle = CreateFile(file_name, GENERIC_READ, FILE_SHARE_READ, nullptr, OPEN_EXISTING, 0, nullptr);
 
     VERIFY(m_file_handle != INVALID_HANDLE_VALUE);
-    const auto file_size = static_cast<u32>(GetFileSize(m_file_handle, nullptr));
+    const auto file_size = static_cast<size_t>(GetFileSize(m_file_handle, nullptr));
     
     const auto file_mapping_handle = CreateFileMapping(m_file_handle, nullptr, PAGE_READONLY, 0, 0, nullptr);
     VERIFY(file_mapping_handle != INVALID_HANDLE_VALUE);
@@ -24,7 +24,7 @@ void CFileStreamReader::construct(pcstr file_name, const u32& window_size)
     VERIFY(m_file_handle != -1);
     struct stat file_info;
     ::fstat(m_file_handle, &file_info);
-    u32 file_size = (u32)file_info.st_size;
+    size_t file_size = (size_t)file_info.st_size;
     inherited::construct(m_file_handle, 0, file_size, file_size, window_size);
 #endif
 }
