+++ src/xrCore/stream_reader_inline.h
@@ -15,25 +15,25 @@ IC void CStreamReader::unmap() { UnmapViewOfFile(m_current_map_view_of_file); }
 #else
 IC void CStreamReader::unmap() { ::munmap(const_cast<u8*>(m_current_map_view_of_file), m_current_window_size); }
 #endif
-IC void CStreamReader::remap(const u32& new_offset)
+IC void CStreamReader::remap(const size_t& new_offset)
 {
     unmap();
     map(new_offset);
 }
 
-IC u32 CStreamReader::elapsed() const
+IC size_t CStreamReader::elapsed() const
 {
-    const u32 offset_from_file_start = tell();
+    const size_t offset_from_file_start = tell();
     VERIFY(m_file_size >= offset_from_file_start);
     return (m_file_size - offset_from_file_start);
 }
 
-IC const u32& CStreamReader::length() const { return m_file_size; }
+IC const size_t& CStreamReader::length() const { return m_file_size; }
 IC void CStreamReader::seek(const int& offset) { advance(offset - tell()); }
-IC u32 CStreamReader::tell() const
+IC size_t CStreamReader::tell() const
 {
     VERIFY(m_current_pointer >= m_start_pointer);
-    VERIFY(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
+    VERIFY(size_t(m_current_pointer - m_start_pointer) <= m_current_window_size);
     return (m_current_offset_from_start + (m_current_pointer - m_start_pointer));
 }
