+++ src/xrEngine/xrTheora_Stream.cpp
@@ -48,7 +48,7 @@ void CTheoraStream::Reset()
 int CTheoraStream::ReadData()
 {
     char* buffer = ogg_sync_buffer(&o_sync_state, 4096);
-    long bytes = 4096 > (size_t)source->elapsed() ? source->elapsed() : 4096;
+    size_t bytes = 4096 > source->elapsed() ? source->elapsed() : 4096;
     source->r(buffer, bytes);
     ogg_sync_wrote(&o_sync_state, bytes);
     return bytes;
